Stop client from looping without time delay when streaming DNS lookup fails (#164)
When curl returned an error representing failure to lookup host, the client
would instantly loop around and try again.

This was because:
1. The error wasn't triggering any custom logic, such as the "permanent error" logic, to exit the loop
2. Therefore it was only checking if intentionallyClosed was set.

intentionallyClosed is problematic. It is supposed (?) to represent the fact that
the stream was intentionally closed, for instance due to reinitialization. In that case,
it would be set to 0.

In this bug, it is set to 0 not because it was intentionally closed, but because
curl doesn't initialize the socket at all if DNS lookup fails.

The fix is to treat the stream handle as somewhat of a tribool, and have a separate meaning for
"uninitialized" vs "intentionallyClosed". This is encapsulated in the new socket.h API.