Switch Releaser linux job to docker executor; add empty contexts to Mac & Windows jobs (#189)
The purpose of this change is to fix the Releaser config, which was recently broken by changes in
releaser that introduced the usage of the aws-cli orb.

The change to the Linux job to docker allows that job to "just work", because the aws commands are run
directly in the container. I don't think there's a particular reason to use a circleCI linux job rather than
running this one in Docker.

The addition of 'context: circleci-oidc-auth' to the Mac and Windows jobs is necessary for the aws commands to succeed.