* fix: update LDi_castJSONToValue signature to eliminate GCC 11 compiler error
The LDi_castJSONToValue function is responsible for obtaining a concrete value from an LDJSON,
given a type specified at runtime. For example, it maps LDBool (an enum value) to the LDGetBool
function, which returns an LDBoolean.

In this way, all the LDVariation methods can share a common implementation  for obtaining a flag value,
while still offering static types to users.

The gist of the issue is that GCC 11 thinks it's possible for a codepath to be executed where
the boolean return value (1 byte) from LDBoolVariation would be overwritten with a double (8 bytes). This is because
LDJSONGetType is called twice from LDi_evalInternal:
1) To check that the flag's value type is the same as the one requested by the user (if not, return a default value)
2) To cast the flag's value into the concrete return type.

The problem is that both checks are invoked on (node->flag.value) independently. I believe since GCC cannot prove that
both calls will return the same value, it triggers the warning.

The fix is to compute the type once, and then pass that into the cast function.